backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-vision==3.4.4
